// packages/astro/dist/core/errors/errors-data.d.ts
export declare const ClientAddressNotAvailable: {
	name: string;
	title: string;
	message: (adapterName: string) => string;
};